Move files from internal-api to lib dir
[#165743428]
