+++ docs/lara-plugin-api/interfaces/iembeddableruntimecontext.md
@@ -26,7 +26,7 @@
 
 **● clickToPlayId**: *`string` \| `null`*
 
-*Defined in [types.ts:84](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L84)*
+*Defined in [types.ts:84](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L84)*
 
 DOM id of click to play overlay if enabled.
 
@@ -39,7 +39,7 @@ ___
 
 **● container**: *`HTMLElement`*
 
-*Defined in [types.ts:47](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L47)*
+*Defined in [types.ts:47](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L47)*
 
 Embeddable container.
 
@@ -50,7 +50,7 @@ ___
 
 **● getInteractiveState**: *`function`*
 
-*Defined in [types.ts:68](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L68)*
+*Defined in [types.ts:68](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L68)*
 
 Function that returns interactive state (Promise) or null if embeddable isn't interactive.
 
@@ -66,7 +66,7 @@ ___
 
 **● getReportingUrl**: *`function`*
 
-*Defined in [types.ts:78](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L78)*
+*Defined in [types.ts:78](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L78)*
 
 Function that returns reporting URL (Promise) or null if it's not an interactive or reporting URL is not defined. Note that reporting URL is defined in the interactive state (that can be obtained via #getInteractiveState method). If your code needs both interactive state and reporting URL, you can pass interactiveStatePromise as an argument to this method to limit number of network requests.
 
@@ -90,7 +90,7 @@ ___
 
 **● laraJson**: *`any`*
 
-*Defined in [types.ts:66](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L66)*
+*Defined in [types.ts:66](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L66)*
 
 Serialized form of the embeddable. Defined by LARA export code, so it's format cannot be specified here. Example (interactive):
