+++ docs/lara-plugin-api/interfaces/ioffering.md
@@ -24,7 +24,7 @@
 
 **● id**: *`number`*
 
-*Defined in [types.ts:114](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L114)*
+*Defined in [types.ts:114](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L114)*
 
 ___
 <a id="name"></a>
@@ -33,7 +33,7 @@ ___
 
 **● name**: *`string`*
 
-*Defined in [types.ts:115](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L115)*
+*Defined in [types.ts:115](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L115)*
 
 ___
 <a id="url"></a>
@@ -42,7 +42,7 @@ ___
 
 **● url**: *`string`*
 
-*Defined in [types.ts:116](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/types.ts#L116)*
+*Defined in [types.ts:116](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/types.ts#L116)*
 
 ___
