+++ docs/lara-plugin-api/interfaces/ipopupcontroller.md
@@ -24,7 +24,7 @@
 
 **● close**: *`function`*
 
-*Defined in [popup.ts:43](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/popup.ts#L43)*
+*Defined in [popup.ts:43](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/popup.ts#L43)*
 
 Closes popup (display: none). Also removes HTML element from DOM tree if `removeOnClose` is equal to true.
 
@@ -40,7 +40,7 @@ ___
 
 **● open**: *`function`*
 
-*Defined in [popup.ts:41](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/popup.ts#L41)*
+*Defined in [popup.ts:41](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/popup.ts#L41)*
 
 Opens popup (makes sense only if autoOpen option is set to false during initialization).
 
@@ -56,7 +56,7 @@ ___
 
 **● remove**: *`function`*
 
-*Defined in [popup.ts:45](https://github.com/concord-consortium/lara/blob/d238af32/lara-typescript/src/plugin-api/popup.ts#L45)*
+*Defined in [popup.ts:45](https://github.com/concord-consortium/lara/blob/80a682ff/lara-typescript/src/plugin-api/popup.ts#L45)*
 
 Removes HTML element from DOM tree.
