+++ lara-typescript/src/internal-api/index.ts
@@ -1,1 +1,1 @@
-export { initPlugin } from "./plugins";
+export { initPlugin } from "../lib/plugins";
