+++ lara-typescript/src/lib/embeddable-runtime-context.ts
@@ -1,6 +1,6 @@
 import {
   IEmbeddableRuntimeContext, IInteractiveState
-} from "../plugin-api/";
+} from "../plugin-api";
 
 /** Data generated by LARA, passed to .initPlugin function, and then used to generate final IEmbeddableRuntimeContext */
 export interface IEmbeddableContext {
